ProductItem stylized
