+++ src/App.css
@@ -11,29 +11,6 @@
   width: 363px;
 }
 
-.products-header {
-  text-align: center;
-  line-height: 60%;
-}
-
-.products-grid {
-  display: grid;
-  grid-template-columns: repeat(4, 250px);
-  grid-auto-rows: 200px;
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
 .navigation-bar {
   text-align: center;
-}
-
-.product-item {
-  width: 170px;
-  height: 135px;
-  border: 3px outset gray;
-  padding: 20px;
-  line-height: 50px;
 }
